author modifiedmodified

+++ routes/authors.mjs
@@ -7,15 +7,16 @@ import { Author } from "../mongoose/schemas/author.mjs"
 import { Genre } from "../mongoose/schemas/genre.mjs"
 import { Publisher } from "../mongoose/schemas/publisher.mjs"
 import bodyParser from "body-parser"
-
+import mongoose from "mongoose"
 
 
 const router = Router()
 
 //add new author
 router.post("/api/authors", adminAuth, async (request, response) => {
-    const {firstName, lastName, dateOfBirth, nationality, biography} = request.body
+   
     try {
+        const {firstName, lastName, dateOfBirth, nationality, biography} = request.body
         const myAuthor = await Author.create({firstName, lastName, dateOfBirth, nationality, biography})
         return response.status(201).send(myAuthor)
     } catch (error){
@@ -24,9 +25,13 @@ router.post("/api/authors", adminAuth, async (request, response) => {
 })
 
 //get detail of an author
-router.get("/api/authors/detail", adminAuth, async (request, response) => {
-    const {_id} = request.body
+router.get("/api/authors/:_id", adminAuth, async (request, response) => {
+   
     try {
+        const {_id} = request.params
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return response.status(400).send({ message: 'Invalid ID format' });
+            }
         const queriedAuthor = await Author.findById(_id)
         if (queriedAuthor){
             return response.status(201).send(queriedAuthor)
@@ -39,9 +44,13 @@ router.get("/api/authors/detail", adminAuth, async (request, response) => {
 } )
 
 //delete an author
-router.delete("/api/authors", adminAuth, async (request, response) => {
-    const {_id} = request.body
+router.delete("/api/authors/:_id", adminAuth, async (request, response) => {
+    
     try {
+        const {_id} = request.params
+        if (!mongoose.Types.ObjectId.isValid(_id)) {
+            return response.status(400).send({ message: 'Invalid ID format' });
+            }
         const deletedAuthor = await Author.findByIdAndDelete(_id)
         if (deletedAuthor){
             return response.status(201).send(`Deleted author: ${deletedAuthor}`)
@@ -54,42 +63,35 @@ router.delete("/api/authors", adminAuth, async (request, response) => {
 })
 
 //list authors
-router.get("/api/authors/list", adminAuth, async (request, response) => {
+router.get("/api/authors", adminAuth, async (request, response) => {
     try {
         const authorList = await Author.find().sort({ dateOfBirth : -1 })
-        if (authorList.length > 0){
-            return response.status(201).send(authorList)
-        } else {
-            return response.send("The list is empty")
-        }
+       
+        return response.status(201).send(authorList)
+        
     } catch (error) {
         return response.status(400).send(error.message)
     }
 }) 
 
 //update author
-router.patch("/api/authors", adminAuth, async (request, response) => {
-    let isError = []
-    const {_id} = request.body
-    if (!_id){
-        return response.status(400).send(`id is missing`)
-    }
-
+router.patch("/api/authors/:_id", adminAuth, async (request, response) => {
+    try{
+    const {_id} = request.params
+    if (!mongoose.Types.ObjectId.isValid(_id)) {
+        return response.status(400).send({ message: 'Invalid ID format' });
+        }
     const myAuthor = await Author.findById(_id)
-
+    if (!myAuthor){
+        return response.status(400).send({ message: 'Cannot query document' });
+    }
     Object.entries(request.body).forEach(([key, value]) => {
         if (value){
           myAuthor[key] = value
-        } else {
-          isError.push(key)
-        }
+        } 
       })
-
-    if (isError.length >= 1){
-        return response.status(400).send(`You are not allowed update "${isError}"`)  
-      }
         
-    try {
+
         const updatedAuthor = await myAuthor.save()
         return response.status(201).send(updatedAuthor)
         }
